Move a function
