+++ app/components/UI/Predict/hooks/usePredictToastRegistrations.tsx
@@ -19,11 +19,15 @@ import type { PredictTransactionStatusChangedPayload } from '../controllers/Pred
 import { getEvmAccountFromSelectedAccountGroup } from '../utils/accounts';
 import { formatPrice } from '../utils/format';
 import { predictQueries } from '../queries';
+import type { Hex } from '@metamask/utils';
 import { usePredictClaim } from './usePredictClaim';
 import { usePredictDeposit } from './usePredictDeposit';
 import { usePredictWithdraw } from './usePredictWithdraw';
 import { store } from '../../../../store';
-import { getWithdrawConfirmedMessage } from '../utils/withdrawConfirmedMessage';
+import { selectTransactionMetadataById } from '../../../../selectors/transactionController';
+import { selectSingleTokenByAddressAndChainId } from '../../../../selectors/tokensController';
+import { selectTickerByChainId } from '../../../../selectors/networkController';
+import type { RootState } from '../../../../reducers';
 
 const showPendingToast = ({
   showToast,
@@ -341,3 +345,62 @@ export const usePredictToastRegistrations = (): ToastRegistration[] => {
     [handleTransactionStatusChanged],
   );
 };
+
+/**
+ * Derives the withdraw-confirmed toast message from transaction metadata.
+ *
+ * For post-quote withdrawals with valid metamaskPay data, displays the
+ * resolved token symbol and targetFiat amount. Otherwise falls back to the
+ * original USDC-only message.
+ */
+function getWithdrawConfirmedMessage(
+  state: RootState,
+  transactionId: string | undefined,
+  fallbackAmount: number,
+): { title: string; description: string } {
+  const title = strings('predict.withdraw.withdraw_completed');
+
+  const txMeta = transactionId
+    ? selectTransactionMetadataById(state, transactionId)
+    : undefined;
+  const { metamaskPay } = txMeta ?? {};
+
+  if (!metamaskPay?.isPostQuote) {
+    return {
+      title,
+      description: strings('predict.withdraw.withdraw_completed_subtitle', {
+        amount: formatPrice(fallbackAmount),
+      }),
+    };
+  }
+
+  const targetFiat = Number(metamaskPay.targetFiat);
+  const withdrawAmount =
+    Number.isFinite(targetFiat) && targetFiat > 0 ? targetFiat : fallbackAmount;
+
+  const chainId = metamaskPay.chainId as Hex | undefined;
+  const tokenAddress = metamaskPay.tokenAddress as Hex | undefined;
+
+  let tokenSymbol: string | undefined;
+  if (tokenAddress && chainId) {
+    tokenSymbol = selectSingleTokenByAddressAndChainId(
+      state,
+      tokenAddress,
+      chainId,
+    )?.symbol;
+  }
+  if (!tokenSymbol && chainId) {
+    tokenSymbol = selectTickerByChainId(state, chainId);
+  }
+  if (!tokenSymbol) {
+    tokenSymbol = 'USDC';
+  }
+
+  return {
+    title,
+    description: strings(
+      'predict.withdraw.withdraw_any_token_completed_subtitle',
+      { amount: formatPrice(withdrawAmount), token: tokenSymbol },
+    ),
+  };
+}
